requirements.txt
anyio==3.6.2
astropy==5.2.2
attrs==23.1.0
bitarray==2.7.3
bitstring==4.0.2
boto3==1.26.123
botocore==1.29.123
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
ciso8601==2.3.0
click==8.1.3
colorama==0.4.6
gevent==1.5.0
greenlet==0.4.16
h11==0.14.0
idna==3.4
importlib-resources==5.12.0
influxdb-client==1.36.1
inotify==0.2.10
jmespath==1.0.1
jsonschema==4.17.3
nats-py==2.2.0
nose==1.3.7
numpy==1.24.3
msgpack==1.0.5
packaging==23.1
pandas==2.0.1
pkgutil_resolve_name==1.3.10
portion==2.4.0
pyasn1==0.5.0
pycparser==2.21
pyerfa==2.0.0.3
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
reactivex==4.0.4
requests==2.31.0
s3transfer==0.6.0
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.27.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
websockets==11.0.2
zipp==3.15.0
